pyproject.toml
[project]
name = "vox"
version = "0.1.0"
description = "Open source macOS dictation app"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "rumps>=0.4.0",
    "pynput>=1.7.6",
    "sounddevice>=0.4.6",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "pyobjc-framework-ApplicationServices>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyperclip>=1.8.2",
    "numpy>=1.24.0",
]

[project.scripts]
vox = "vox.app:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-timeout>=2.2.0",
    "pytest-mock>=3.12.0",
    "hypothesis>=6.100.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
timeout = 30
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (may use real audio/model)",
    "e2e: End-to-end tests (full app, agent-orchestrated)",
    "slow: Slow tests (model loading, etc.)",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vox"]
